[UPD] Mokka: Trying a new method to compute the parameters used by the playback timer. Standard playback frequency set every time to 25Hz
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@958 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Tools/Mokka/TimeEventControlerWidget.h
@@ -152,7 +152,7 @@ private:
   QIcon* mp_NextEventIcon;
   QIcon* mp_NextEventActiveIcon;
   QTimer* mp_Timer;
-  int m_PlaybackStep;
+  double m_PlaybackStep;
   int m_PlaybackDelay;
   bool m_OpenEditorWhenInserting;
 };
